+++ pyseto/versions/v2.py
@@ -107,6 +107,9 @@ class V2Public(SodiumKey):
 
         if not isinstance(self._key, (Ed25519PublicKey, Ed25519PrivateKey)):
             raise ValueError("The key is not Ed25519 key.")
+
+        if isinstance(self._key, Ed25519PublicKey):
+            self._is_secret = False
         return
 
     def to_paserk_id(self) -> str: